chore: 타입, import 수정
