+++ components/hospital/super/hospitals/hos-user-sheet.tsx
@@ -11,36 +11,39 @@ import {
   SheetTitle,
   SheetTrigger,
 } from '@/components/ui/sheet'
-import { getStaffs } from '@/lib/services/admin/staff/staff'
-import type { UserHospitalJoined } from '@/types/adimin'
-import { ArrowLeft } from 'lucide-react'
+import { getStaffs, type Staff } from '@/lib/services/admin/staff'
+import { ArrowLeft, LoaderCircle } from 'lucide-react'
 import { useState } from 'react'
 
-export default function HosUserSheet({
-  hosId,
-  hosName,
-}: {
+type Props = {
   hosId: string
   hosName: string
-}) {
+}
+
+export default function HosUserSheet({ hosId, hosName }: Props) {
   const [isSheetOpen, setIsSheetOpen] = useState(false)
-  const [staffs, setStaffs] = useState<UserHospitalJoined[]>([])
+  const [isFetching, setIsFetching] = useState(false)
+  const [staffs, setStaffs] = useState<Staff[]>([])
 
-  const handleSheetOpenChange = async () => {
-    setIsSheetOpen(!isSheetOpen)
-
-    if (!isSheetOpen) {
+  const handleSheetOpenChange = async (open: boolean) => {
+    if (open) {
+      setIsFetching(true)
       const staffs = await getStaffs(hosId)
-
       setStaffs(staffs)
+      setIsFetching(false)
     }
+    setIsSheetOpen(open)
   }
 
   return (
     <Sheet open={isSheetOpen} onOpenChange={handleSheetOpenChange}>
       <SheetTrigger asChild>
-        <Button variant="ghost">
-          <ArrowLeft size={18} />
+        <Button variant="ghost" size="icon">
+          {isFetching ? (
+            <LoaderCircle className="animate-spin" />
+          ) : (
+            <ArrowLeft size={18} />
+          )}
         </Button>
       </SheetTrigger>
       <SheetContent className="max-w-full md:max-w-[1200px]">
