+++ constants/admin/admin-sidebar-items.ts
@@ -1,11 +1,4 @@
-import {
-  Pill,
-  Syringe,
-  TestTubeDiagonal,
-  UserCheck,
-  Users,
-  Utensils,
-} from 'lucide-react'
+import { Syringe, UserCheck, Users } from 'lucide-react'
 
 export const ADMIN_SIDEBAR_ITEMS = [
   {
@@ -29,25 +22,25 @@ export const ADMIN_SIDEBAR_ITEMS = [
     isReady: true,
     isResponsive: false,
   },
-  {
-    name: '약물설정',
-    path: 'drug-settings',
-    icon: Pill,
-    isReady: false,
-    isResponsive: false,
-  },
-  {
-    name: '사료설정',
-    path: 'diet-settings',
-    icon: Utensils,
-    isReady: false,
-    isResponsive: false,
-  },
-  {
-    name: '검사설정',
-    path: 'test-settings',
-    icon: TestTubeDiagonal,
-    isReady: false,
-    isResponsive: false,
-  },
+  // {
+  //   name: '약물설정',
+  //   path: 'drug-settings',
+  //   icon: Pill,
+  //   isReady: false,
+  //   isResponsive: false,
+  // },
+  // {
+  //   name: '사료설정',
+  //   path: 'diet-settings',
+  //   icon: Utensils,
+  //   isReady: false,
+  //   isResponsive: false,
+  // },
+  // {
+  //   name: '검사설정',
+  //   path: 'test-settings',
+  //   icon: TestTubeDiagonal,
+  //   isReady: false,
+  //   isResponsive: false,
+  // },
 ] as const
